refactor(export): split export.ts into native + web variants with shared html helper
Move `htmlToMarkdown` + `wrapHtmlForPdf` into `utils/html.ts` (pure,
platform-agnostic). Split the platform-dependent IO into `utils/export.ts`
(native: `expo-print`, `expo-sharing`, `expo-file-system` imported
statically) and `utils/export.web.ts` (web: `jspdf`, `html2canvas`, Tauri
dialogs via `@tauri-apps/api`). Removes per-call dynamic `await import(...)`
used as platform gates; Tauri detection via `isTauri()` remains legitimate
inside the web variant.

Part of LL-014. Refs #117
